+++ src/components/Home/Home.css
@@ -1,0 +1,12 @@
+@import url("https://fonts.googleapis.com/css2?family=Lato:ital,wght@0,400;0,700;1,400&display=swap");
+.App {
+  text-align: center;
+  font-family: "Lato", sans-serif;
+}
+
+.search-results {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  padding: 0 12px;
+}
